Refactor toolbar
- related #80
